% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_explicit_founder_parents.R
\name{add_explicit_founder_parents}
\alias{add_explicit_founder_parents}
\title{explicitly add parents of 0 to the founders of a pedigree}
\usage{
add_explicit_founder_parents(P)
}
\arguments{
\item{P}{pedigree with columns Kid Pa Ma}
}
\description{
This seems to be required by Mendel.
}
