% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pedigree_sex_column.R
\name{add_pedigree_sex_column}
\alias{add_pedigree_sex_column}
\title{given a pedigree as Kid Pa Ma, this adds a column called Sex}
\usage{
add_pedigree_sex_column(P, fem_prob = 0.5)
}
\arguments{
\item{P}{the pedigree with columns Kid Pa Ma}

\item{fem_prob}{prob that unknown sex indiv is assigned as female.}
}
\description{
This just uses the occurrence of each each individual in the
pedigree as Ma or Pa to figure out its sex.  Individuals with
no offspring are randomly assigned to be females (2) with probability
fem_prob, and are males (1) otherwise.
This function does not check for sex inconsistencies.
}
