% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biallelic_geno_error.R
\name{biallelic_geno_error}
\alias{biallelic_geno_error}
\title{Simulate genotyping error on biallelic markers}
\usage{
biallelic_geno_error(genos, het_miscall = 0.02, hom_miscall = 0.005)
}
\arguments{
\item{genos}{a vector of genotypes. The genotypes \emph{must} be given as \code{1/1}, \code{1/2}, \code{2/1} or \code{2/2}.}

\item{het_miscall}{the rate at which heterozygotes are called as
homozygotes.}

\item{hom_miscall}{the rate at which a homozygote is called as a het.}
}
\description{
Simple function to simulate genotyping errors.
}
