% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinook_chromosomes}
\alias{chinook_chromosomes}
\title{Lengths of the 34 named Chinook salmon chromosomes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 3 columns.
}
\usage{
chinook_chromosomes
}
\description{
Pretty self-explanatory tibble.
}
\keyword{datasets}
