% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_acyclic_pedigree.R
\name{create_founders}
\alias{create_founders}
\title{quick function to create Nm males and Nf female founders.
This also initializes their connected components, etc.
And, it returns them in a list, like what we shall use for
going forward...}
\usage{
create_founders(Nm = 50, Nf = Nm)
}
\description{
quick function to create Nm males and Nf female founders.
This also initializes their connected components, etc.
And, it returns them in a list, like what we shall use for
going forward...
}
\keyword{internal}
