% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_acyclic_pedigree.R
\name{expand_offspring}
\alias{expand_offspring}
\title{simple function to expand pairs into named individuals
with sex given randomly at 50/50}
\usage{
expand_offspring(P, t)
}
\arguments{
\item{P}{the tibble that comes out of num_offspring}

\item{t}{the year that the newborns are being born}
}
\description{
simple function to expand pairs into named individuals
with sex given randomly at 50/50
}
\keyword{internal}
