% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FourGen_555_pedigree}
\alias{FourGen_555_pedigree}
\title{A four generation pedigree simulated to be acyclic}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 414 rows and 4 columns.
}
\usage{
FourGen_555_pedigree
}
\description{
Eric Anderson simulated this using code in one of the vignettes.
}
\keyword{datasets}
