% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_acyclic_pedigree.R
\name{iteratively_expand_conn_comps}
\alias{iteratively_expand_conn_comps}
\title{function that iteratively expands the connected component members
of each individual, until they reflect all the individuals they
are connected to.}
\usage{
iteratively_expand_conn_comps(P, CC = NULL)
}
\arguments{
\item{P}{a tibble like that which comes out of expand_offspring, or
NULL if you just want to operate on the named list of conn comps.}
}
\description{
function that iteratively expands the connected component members
of each individual, until they reflect all the individuals they
are connected to.
}
\keyword{internal}
