% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_markers}
\alias{long_markers}
\title{A tibble of a bunch of multiallelic markers with allele frequencies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 53637 rows and 7 columns.
}
\usage{
long_markers
}
\description{
It's a tibble
}
\keyword{datasets}
