% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_acyclic_pedigree.R
\name{make_pairs}
\alias{make_pairs}
\title{Determine pairs.  Given the individuals in a given year
determine the pairs that will be formed, add in founders
as necessary to make pairings.}
\usage{
make_pairs(L, t)
}
\arguments{
\item{L}{a list with elements "males" and "females" which are named
lists of connected components.}

\item{t}{the year that the pairs that are being formed were born.}
}
\description{
Determine pairs.  Given the individuals in a given year
determine the pairs that will be formed, add in founders
as necessary to make pairings.
}
\keyword{internal}
