% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_acyclic_pedigree.R
\name{num_offspring}
\alias{num_offspring}
\title{this little function takes the tibble that comes out of
make_pairs, and tells us how many offspring each will have.
For now, I am setting it to that multinom + 1, but that can
easily change in the future.}
\usage{
num_offspring(P)
}
\arguments{
\item{P}{the tibble that comes out of make_pairs}
}
\description{
this little function takes the tibble that comes out of
make_pairs, and tells us how many offspring each will have.
For now, I am setting it to that multinom + 1, but that can
easily change in the future.
}
\keyword{internal}
