% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ped5gen_acyclic}
\alias{ped5gen_acyclic}
\title{A five generation pedigree simulated to be acyclic}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 600 rows and 4 columns.
}
\usage{
ped5gen_acyclic
}
\description{
Eric Anderson simulated this.
}
\keyword{datasets}
