% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedigrees}
\alias{pedigrees}
\title{A list of some very small pedigrees used in CKMRsim}
\format{
An object of class \code{list} of length 11.
}
\usage{
pedigrees
}
\description{
I just have these here for some examples.
}
\keyword{datasets}
