% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_acyclic_pedigree.R
\name{set_init_cc_members}
\alias{set_init_cc_members}
\title{set the initial connected component members of all newborns.
This will be the connected components of the parents plus
all of their siblings.}
\usage{
set_init_cc_members(P)
}
\arguments{
\item{P}{a tibble like what comes out of expand_offspring}

\item{I}{the Indivs list component for the parent (t-1) generation}
}
\description{
set the initial connected component members of all newborns.
This will be the connected components of the parents plus
all of their siblings.
}
\keyword{internal}
