% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_acyclic_pedigree.R
\name{translate_tib_to_list_element}
\alias{translate_tib_to_list_element}
\title{this function takes the tibble coming out of
iteratively _expand_conn_comps, and it turn it
into a list item that can be returned into Indiv[\link{t}].}
\usage{
translate_tib_to_list_element(Pairs)
}
\description{
this function takes the tibble coming out of
iteratively _expand_conn_comps, and it turn it
into a list item that can be returned into Indiv[\link{t}].
}
\keyword{internal}
