% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_acyclic_pedigree.R
\name{update_conn_comps_after_pair_union}
\alias{update_conn_comps_after_pair_union}
\title{update the conn-comps of two mates and all the rest.
This assumes that the current connected components are in two
named lists, Mal and Fem.  It updates each, and then smushes
them into a single list and iteratively updates the remaining
ones, and then returns the updated Mal and Fem in a list.
mn and fn are the names of the male and female that are paired up.}
\usage{
update_conn_comps_after_pair_union(Mal, Fem, mn, fn)
}
\description{
update the conn-comps of two mates and all the rest.
This assumes that the current connected components are in two
named lists, Mal and Fem.  It updates each, and then smushes
them into a single list and iteratively updates the remaining
ones, and then returns the updated Mal and Fem in a list.
mn and fn are the names of the male and female that are paired up.
}
\keyword{internal}
